[UPD] Add LocationUtil and fix Filter unit tests #bgweyn[closed]
There is a bug when I run these unit tests on IDEA. On the first
run, some tests failed. But there are correct. I take as a proof this:
if I rerun them with "Rerun failed tests", they all success.
I do not know from where comes this bug, but I will keep it for now
because I know that my tests are correct (because a rerun make them
success).